feat[fzs79-77-75]: Implementando backend da parte da página de folga

+++ src/main/java/com/fromzero/checkpoint/repositories/ColaboradorRepository.java
@@ -1,13 +1,9 @@
 package com.fromzero.checkpoint.repositories;
 
-import com.fromzero.checkpoint.models.Colaborador;
+import com.fromzero.checkpoint.entities.Colaborador;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ColaboradorRepository extends JpaRepository<Colaborador, Integer> {
-    // Você pode adicionar métodos de consulta personalizados aqui, se necessário
-    // Por exemplo, para buscar um colaborador pelo email:
-    // Optional<Colaborador> findByEmail(String email);
+public interface ColaboradorRepository extends JpaRepository<Colaborador, Long> {
 }
-
